changed datatype again to int

+++ src/main/java/com/example/ABCBankDetails/dto/PayeeAddRequestDTO.java
@@ -8,20 +8,20 @@ public class PayeeAddRequestDTO implements Serializable {
 	
 	private static final long serialVersionUID = 1L;
 	
-	private Long customerId;
-	private Long payeeId;
+	private int customerId;
+	private int payeeId;
 	
 	
 	public Long getCustomerId() {
 		return customerId;
 	}
-	public void setCustomerId(Long customerId) {
+	public void setCustomerId(int customerId) {
 		this.customerId = customerId;
 	}
 	public Long getPayeeId() {
 		return payeeId;
 	}
-	public void setPayeeId(Long payeeId) {
+	public void setPayeeId(int payeeId) {
 		this.payeeId = payeeId;
 	}
